bayo gun charge bugfix
